+++ src/pages/booklet/biodata.html
@@ -1,0 +1,1 @@
+<html> <head> <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0"> <title>Biodata Booklet 3'2021</title> <style type="text/css"> html{ margin: 0; height: 100%; overflow: hidden; } iframe{ position: absolute; left:0; right:0; bottom:0; top:0; border:0; } </style> </head> <body> <iframe id="typeform-full" width="100%" height="100%" frameborder="0" allow="camera; microphone; autoplay; encrypted-media;" src="https://form.typeform.com/to/QNDbPUfv"></iframe> <script type="text/javascript" src="https://embed.typeform.com/embed.js"></script> </body> </html>
